+++ extensions/manotori/src/utils.ts
@@ -3,5 +3,6 @@ import { ErrorResponse } from "./types";
 export async function parseResponse(response: Response) {
   const res = await response.json();
   if (!response.ok) throw new Error((res as ErrorResponse).msg);
-  return res;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  return res as any;
 }