+++ extensions/manotori/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Manotori Changelog
 
+## [View DNS Zones, DNS Records] - {PR_MERGE_DATE}
+
+- View DNS Zones
+- View DNS Records
+    - Create DNS Record
+
 ## [View NS Sets] - 2025-01-03
 
 - View `Nameserver Sets`